06 Checkered 2D

+++ raytracer/raytracer/math/functions/checkered2d.h
@@ -7,6 +7,6 @@ namespace math
 {
 	namespace functions
 	{
-		Function<bool(const Point2D&)> checkered2d(int size = 1);
+		Function<bool(const Point2D&)> checkered2d(double size = 1);
 	}
 }
